requirements.txt
Flask==2.0.1
cfenv==0.5.3
sap-xssec==3.0.0
Shapely==1.8.0
seaborn==0.11.2